Refactored out component for use by PaMessagesPage.
