+++ assets/css/screen-detail-action-bar.scss
@@ -23,7 +23,10 @@
   }
 }
 
-.screen-detail-action-bar-button {
+.three-dots-vertical-button {
+  width: 36px;
+  margin-left: 8px;
+  float: right;
   border-radius: 6px;
   background-color: $gray-700;
   border: 1px solid rgba(23, 31, 38, 0.3);
@@ -38,12 +41,6 @@
     outline: none;
     box-shadow: none;
   }
-}
-
-.three-dots-vertical-button {
-  width: 36px;
-  margin-left: 8px;
-  float: right;
 
   &__icon {
     position: relative;
